@@ -34,7 +34,7 @@
   },
   "dependencies": {
     "cluster-key-slot": "^1.1.0",
-    "debug": "^4.3.1",
+    "debug": "^4.3.3",
     "denque": "^1.1.0",
     "lodash.defaults": "^4.2.0",
     "lodash.flatten": "^4.4.0",
@@ -48,39 +48,36 @@
   "devDependencies": {
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
-    "@types/bluebird": "^3.5.30",
     "@types/chai": "^4.2.11",
     "@types/chai-as-promised": "^7.1.3",
     "@types/debug": "^4.1.5",
     "@types/lodash.defaults": "^4.2.6",
     "@types/lodash.flatten": "^4.4.6",
     "@types/lodash.isarguments": "^3.1.6",
-    "@types/mocha": "^7.0.2",
+    "@types/mocha": "^8.2.3",
     "@types/node": "^13.11.0",
     "@types/redis-errors": "1.2.0",
     "@types/sinon": "^9.0.0",
-    "@types/uuid": "^8.3.0",
-    "@typescript-eslint/eslint-plugin": "^1.13.0",
-    "@typescript-eslint/parser": "^2.26.0",
-    "bluebird": "^3.7.2",
+    "@typescript-eslint/eslint-plugin": "^5.10.2",
+    "@typescript-eslint/parser": "^5.10.2",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "cronometro": "^0.6.0",
     "cz-conventional-changelog": "^3.1.0",
-    "eslint": "^5.16.0",
-    "eslint-config-prettier": "^6.10.1",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
     "husky": "^4.2.3",
-    "mocha": "^6.2.3",
-    "prettier": "^2.0.2",
+    "mocha": "^8.4.0",
+    "prettier": "^2.5.1",
     "pretty-quick": "^2.0.1",
     "server-destroy": "^1.0.1",
     "sinon": "^9.0.1",
-    "ts-node": "^8.8.1",
-    "typescript": "3.8.3",
+    "ts-node": "^10.4.0",
+    "type-fest": "^2.11.1",
+    "typescript": "^4.5.5",
     "uuid": "^8.3.0"
   },
   "engines": {
-    "node": ">=6"
+    "node": ">=10"
   },
   "config": {
     "commitizen": {
