@@ -22,11 +22,11 @@ describe("cluster:tls option", () => {
     new MockServer(30002, argvHandler);
     new MockServer(30003, argvHandler);
 
-    // @ts-ignore
+    // @ts-expect-error
     const stub = sinon.stub(tls, "connect").callsFake((op) => {
-      // @ts-ignore
+      // @ts-expect-error
       expect(op.ca).to.eql("123");
-      // @ts-ignore
+      // @ts-expect-error
       expect(op.port).to.be.oneOf([30001, 30003, 30003]);
       const stream = net.createConnection(op);
       stream.on("connect", (data) => {
